requirements.txt
colorama==0.4.4
commonmark==0.9.1
paho-mqtt==1.6.1
Pygments==2.11.2
rich==11.0.0
